9C-1141-2 Collections APIAppsInfoRequest
We extract from the Nine Cards API a set of routes related to the
`collections` routes, which deal with endpoints and functionalities
for shared collections, subscriptions, etc.
