+++ modules/api/src/main/scala/cards/nine/api/JsonFormats.scala
@@ -1,52 +1,16 @@
 package cards.nine.api
 
-import cats.data.Xor
 import cards.nine.api.messages.GooglePlayMessages._
 import cards.nine.api.messages.InstallationsMessages._
-import cards.nine.api.messages.SharedCollectionMessages._
 import cards.nine.api.messages.UserMessages._
 import cards.nine.domain.application.{ Package, Widget }
 import cards.nine.domain.account._
-import cards.nine.processes.collections.messages._
-import io.circe.{ Decoder, Encoder, Json }
-import org.joda.time.DateTime
-import org.joda.time.format.DateTimeFormat
 import spray.httpx.SprayJsonSupport
 import spray.json._
 
 trait JsonFormats
   extends DefaultJsonProtocol
   with SprayJsonSupport {
-
-  implicit object JodaDateTimeFormat extends RootJsonFormat[DateTime] {
-    val formatter = DateTimeFormat.forPattern("yyyy-MM-dd'T'HH:mm:ss.SSSZ").withZoneUTC
-    val dateExample = formatter.print(0)
-
-    def error(v: String) = deserializationError(
-      s"'$v' is not a valid date value. The format for dates must be: '$dateExample'"
-    )
-
-    val decodeDateTime: Decoder[DateTime] = Decoder.instance { cursor ⇒
-      cursor.as[String].flatMap {
-        dateTime ⇒ Xor.right(DateTime.parse(dateTime, formatter))
-      }
-    }
-
-    val encodeDateTime: Encoder[DateTime] = Encoder.instance { dateTime: DateTime ⇒
-      Json.fromString(formatter.print(dateTime))
-    }
-
-    def write(obj: DateTime): JsValue = encodeDateTime(obj).as[String].fold(
-      f ⇒ serializationError(f.message),
-      v ⇒ JsString(v)
-    )
-
-    def read(json: JsValue): DateTime = json match {
-      case JsString(s) ⇒ decodeDateTime(Json.fromString(s).hcursor).fold(_ ⇒ error(s), d ⇒ d)
-      case _ ⇒ error(json.toString)
-    }
-
-  }
 
   implicit object PackageJsonFormat extends JsonFormat[Package] {
     def read(json: JsValue): Package = Package(StringJsonFormat.read(json))
@@ -91,28 +55,6 @@ trait JsonFormats
 
   implicit val updateInstallationResponseFormat = jsonFormat2(ApiUpdateInstallationResponse)
 
-  implicit val appInfoFormat = jsonFormat7(ApiCollectionApp)
-
-  implicit val apiSharedCollection = jsonFormat12(ApiSharedCollection)
-
-  implicit val apiSharedCollectionList = jsonFormat1(ApiSharedCollectionList)
-
-  implicit val apiCreateCollectionRequestFormat = jsonFormat8(ApiCreateCollectionRequest)
-
-  implicit val packagesStatsFormat = jsonFormat2(PackagesStats)
-
-  implicit val apiCreateCollectionResponseFormat = jsonFormat2(ApiCreateOrUpdateCollectionResponse)
-
-  implicit val apiIncreaseViewsCountByOneResponseFormat = jsonFormat1(ApiIncreaseViewsCountByOneResponse)
-
-  implicit val apiSubscribeResponseFormat = jsonFormat0(ApiSubscribeResponse)
-
-  implicit val apiUnsubscribeResponseFormat = jsonFormat0(ApiUnsubscribeResponse)
-
-  implicit val sharedCollectionUpdateInfoFormat = jsonFormat1(SharedCollectionUpdateInfo)
-
-  implicit val apiUpdateCollectionRequestFormat = jsonFormat2(ApiUpdateCollectionRequest)
-
   implicit val apiCategorizeAppsRequestFormat = jsonFormat1(ApiAppsInfoRequest)
 
   implicit val apiCategorizedAppFormat = jsonFormat2(ApiCategorizedApp)
@@ -126,8 +68,6 @@ trait JsonFormats
 
   implicit val apiSetAppInfoRequestFormat = jsonFormat7(ApiSetAppInfoRequest)
   implicit val apiSetAppInfoResponseFormat = jsonFormat0(ApiSetAppInfoResponse)
-
-  implicit val apiGetSubscriptionsByUserResponseFormat = jsonFormat1(ApiGetSubscriptionsByUser)
 
   implicit val apiGetRecommendationsByCategoryRequestFormat = jsonFormat2(ApiGetRecommendationsByCategoryRequest)
