+++ modules/api/src/main/scala/cards/nine/api/NineCardsApi.scala
@@ -3,10 +3,10 @@ package cards.nine.api
 import akka.actor.ActorRefFactory
 import cards.nine.api.NineCardsDirectives._
 import cards.nine.api.NineCardsHeaders.Domain._
+import cards.nine.api.collections.CollectionsApi
 import cards.nine.api.converters.Converters._
 import cards.nine.api.messages.GooglePlayMessages._
 import cards.nine.api.messages.InstallationsMessages._
-import cards.nine.api.messages.SharedCollectionMessages._
 import cards.nine.api.messages.UserMessages._
 import cards.nine.api.utils.SprayMarshallers._
 import cards.nine.api.utils.SprayMatchers._
@@ -15,13 +15,10 @@ import cards.nine.commons.config.Domain.NineCardsConfiguration
 import cards.nine.domain.account.SessionToken
 import cards.nine.domain.analytics._
 import cards.nine.domain.application.{ BasicCard, Category, FullCard, Package, PriceFilter }
-import cards.nine.domain.pagination.Page
 import cards.nine.processes._
-import cards.nine.processes.collections.SharedCollectionProcesses
 import cards.nine.processes.NineCardsServices._
 
 import scala.concurrent.ExecutionContext
-import spray.http.StatusCodes.NotFound
 import spray.routing._
 
 class NineCardsRoutes(
@@ -32,7 +29,6 @@ class NineCardsRoutes(
   applicationProcesses: ApplicationProcesses[NineCardsServices],
   rankingProcesses: RankingProcesses[NineCardsServices],
   recommendationsProcesses: RecommendationsProcesses[NineCardsServices],
-  sharedCollectionProcesses: SharedCollectionProcesses[NineCardsServices],
   refFactory: ActorRefFactory,
   executionContext: ExecutionContext
 ) {
@@ -40,19 +36,19 @@ class NineCardsRoutes(
   import Directives._
   import JsonFormats._
 
-  val nineCardsRoutes: Route = pathPrefix(Segment) {
-    case "apiDocs" ⇒ swaggerRoute
-    case "collections" ⇒ sharedCollectionsRoute
-    case "applications" ⇒ applicationRoute
-    case "recommendations" ⇒ recommendationsRoute
-    case "installations" ⇒ installationsRoute
-    case "login" ⇒ userRoute
-    case "rankings" ⇒ rankings.route
-    case "widgets" ⇒ widgetRoute
-    case _ ⇒ complete(NotFound)
-  }
+  lazy val nineCardsRoutes: Route = {
+    val applicationRoute = new ApplicationRoutes().route
+    val collectionsRoute = new CollectionsApi().route
 
-  private[this] val applicationRoute = new ApplicationRoutes().route
+    pathPrefix("apiDocs")(swaggerRoute) ~
+      pathPrefix("applications")(applicationRoute) ~
+      collectionsRoute ~
+      pathPrefix("installations")(installationsRoute) ~
+      pathPrefix("login")(userRoute) ~
+      pathPrefix("rankings")(rankings.route) ~
+      pathPrefix("recommendations")(recommendationsRoute) ~
+      pathPrefix("widgets")(widgetRoute)
+  }
 
   private[this] lazy val userRoute: Route =
     pathEndOrSingleSlash {
@@ -108,82 +104,6 @@ class NineCardsRoutes(
         }
     }
 
-  private[this] lazy val sharedCollectionsRoute: Route =
-    nineCardsDirectives.authenticateUser { userContext: UserContext ⇒
-      pathEndOrSingleSlash {
-        post {
-          entity(as[ApiCreateCollectionRequest]) { request ⇒
-            nineCardsDirectives.generateNewCollectionInfo { collectionInfo: NewSharedCollectionInfo ⇒
-              complete(createCollection(request, collectionInfo, userContext))
-            }
-          }
-        } ~
-          get {
-            nineCardsDirectives.googlePlayInfo { googlePlayContext ⇒
-              complete(getPublishedCollections(googlePlayContext, userContext))
-            }
-          }
-      } ~
-        (path("latest" / CategorySegment / TypedIntSegment[PageNumber] / TypedIntSegment[PageSize]) & get) {
-          (category: Category, pageNumber: PageNumber, pageSize: PageSize) ⇒
-            nineCardsDirectives.googlePlayInfo { googlePlayContext ⇒
-              complete {
-                getLatestCollectionsByCategory(
-                  category          = category,
-                  googlePlayContext = googlePlayContext,
-                  userContext       = userContext,
-                  pageNumber        = pageNumber,
-                  pageSize          = pageSize
-                )
-              }
-            }
-        } ~
-        (path("top" / CategorySegment / TypedIntSegment[PageNumber] / TypedIntSegment[PageSize]) & get) {
-          (category: Category, pageNumber: PageNumber, pageSize: PageSize) ⇒
-            nineCardsDirectives.googlePlayInfo { googlePlayContext ⇒
-              complete {
-                getTopCollectionsByCategory(
-                  category          = category,
-                  googlePlayContext = googlePlayContext,
-                  userContext       = userContext,
-                  pageNumber        = pageNumber,
-                  pageSize          = pageSize
-                )
-              }
-            }
-        } ~
-        pathPrefix("subscriptions") {
-          pathEndOrSingleSlash {
-            get {
-              complete(getSubscriptionsByUser(userContext))
-            }
-          } ~
-            path(TypedSegment[PublicIdentifier]) { publicIdentifier ⇒
-              put(complete(subscribe(publicIdentifier, userContext))) ~
-                delete(complete(unsubscribe(publicIdentifier, userContext)))
-            }
-        } ~
-        pathPrefix(TypedSegment[PublicIdentifier]) { publicIdentifier ⇒
-          pathEndOrSingleSlash {
-            get {
-              nineCardsDirectives.googlePlayInfo { googlePlayContext ⇒
-                complete(getCollection(publicIdentifier, googlePlayContext, userContext))
-              }
-            } ~
-              put {
-                entity(as[ApiUpdateCollectionRequest]) { request ⇒
-                  complete(updateCollection(publicIdentifier, request))
-                }
-              }
-          } ~
-            path("views") {
-              post {
-                complete(increaseViewsCountByOne(publicIdentifier))
-              }
-            }
-        }
-    }
-
   private[this] lazy val swaggerRoute: Route =
     // This path prefix grants access to the Swagger documentation.
     // Both /apiDocs/ and /apiDocs/index.html are valid paths to load Swagger-UI.
@@ -210,106 +130,6 @@ class NineCardsRoutes(
     userProcesses
       .updateInstallation(toUpdateInstallationRequest(request, userContext))
       .map(toApiUpdateInstallationResponse)
-
-  private[this] def getCollection(
-    publicId: PublicIdentifier,
-    googlePlayContext: GooglePlayContext,
-    userContext: UserContext
-  ): NineCardsService[NineCardsServices, ApiSharedCollection] =
-    sharedCollectionProcesses
-      .getCollectionByPublicIdentifier(
-        userId           = userContext.userId.value,
-        publicIdentifier = publicId.value,
-        marketAuth       = toMarketAuth(googlePlayContext, userContext)
-      )
-      .map(r ⇒ toApiSharedCollection(r.data)(toApiCollectionApp))
-
-  private[this] def createCollection(
-    request: ApiCreateCollectionRequest,
-    collectionInfo: NewSharedCollectionInfo,
-    userContext: UserContext
-  ): NineCardsService[NineCardsServices, ApiCreateOrUpdateCollectionResponse] =
-    sharedCollectionProcesses
-      .createCollection(toCreateCollectionRequest(request, collectionInfo, userContext))
-      .map(toApiCreateOrUpdateCollectionResponse)
-
-  private[this] def increaseViewsCountByOne(
-    publicId: PublicIdentifier
-  ): NineCardsService[NineCardsServices, ApiIncreaseViewsCountByOneResponse] =
-    sharedCollectionProcesses
-      .increaseViewsCountByOne(publicId.value)
-      .map(toApiIncreaseViewsCountByOneResponse)
-
-  private[this] def subscribe(
-    publicId: PublicIdentifier,
-    userContext: UserContext
-  ): NineCardsService[NineCardsServices, ApiSubscribeResponse] =
-    sharedCollectionProcesses
-      .subscribe(publicId.value, userContext.userId.value)
-      .map(toApiSubscribeResponse)
-
-  private[this] def updateCollection(
-    publicId: PublicIdentifier,
-    request: ApiUpdateCollectionRequest
-  ): NineCardsService[NineCardsServices, ApiCreateOrUpdateCollectionResponse] =
-    sharedCollectionProcesses
-      .updateCollection(publicId.value, request.collectionInfo, request.packages)
-      .map(toApiCreateOrUpdateCollectionResponse)
-
-  private[this] def unsubscribe(
-    publicId: PublicIdentifier,
-    userContext: UserContext
-  ): NineCardsService[NineCardsServices, ApiUnsubscribeResponse] =
-    sharedCollectionProcesses
-      .unsubscribe(publicId.value, userContext.userId.value)
-      .map(toApiUnsubscribeResponse)
-
-  private[this] def getLatestCollectionsByCategory(
-    category: Category,
-    googlePlayContext: GooglePlayContext,
-    userContext: UserContext,
-    pageNumber: PageNumber,
-    pageSize: PageSize
-  ): NineCardsService[NineCardsServices, ApiSharedCollectionList] =
-    sharedCollectionProcesses
-      .getLatestCollectionsByCategory(
-        userId     = userContext.userId.value,
-        category   = category.entryName,
-        marketAuth = toMarketAuth(googlePlayContext, userContext),
-        pageParams = Page(pageNumber.value, pageSize.value)
-      )
-      .map(toApiSharedCollectionList)
-
-  private[this] def getPublishedCollections(
-    googlePlayContext: GooglePlayContext,
-    userContext: UserContext
-  ): NineCardsService[NineCardsServices, ApiSharedCollectionList] =
-    sharedCollectionProcesses
-      .getPublishedCollections(userContext.userId.value, toMarketAuth(googlePlayContext, userContext))
-      .map(toApiSharedCollectionList)
-
-  private[this] def getSubscriptionsByUser(
-    userContext: UserContext
-  ): NineCardsService[NineCardsServices, ApiGetSubscriptionsByUser] =
-    sharedCollectionProcesses
-      .getSubscriptionsByUser(userContext.userId.value)
-      .map(toApiGetSubscriptionsByUser)
-
-  private[this] def getTopCollectionsByCategory(
-    category: Category,
-    googlePlayContext: GooglePlayContext,
-    userContext: UserContext,
-    pageNumber: PageNumber,
-    pageSize: PageSize
-  ): NineCardsService[NineCardsServices, ApiSharedCollectionList] =
-    sharedCollectionProcesses
-      .getTopCollectionsByCategory(
-        userId     = userContext.userId.value,
-        category   = category.entryName,
-        marketAuth = toMarketAuth(googlePlayContext, userContext),
-        pageParams = Page(pageNumber.value, pageSize.value)
-      )
-      .map(toApiSharedCollectionList)
 
   private[this] def getRecommendationsByCategory(
     request: ApiGetRecommendationsByCategoryRequest,
