+++ modules/api/src/main/scala/cards/nine/api/converters/Converters.scala
@@ -3,14 +3,12 @@ package cards.nine.api.converters
 import cards.nine.api.NineCardsHeaders.Domain._
 import cards.nine.api.messages.GooglePlayMessages._
 import cards.nine.api.messages.InstallationsMessages._
-import cards.nine.api.messages.SharedCollectionMessages._
 import cards.nine.api.messages.UserMessages._
 import cards.nine.commons.NineCardsService.Result
 import cards.nine.domain.account._
 import cards.nine.domain.analytics.{ RankedAppsByCategory, RankedWidgetsByMoment }
 import cards.nine.domain.application._
 import cards.nine.domain.market.MarketCredentials
-import cards.nine.processes.collections.messages._
 import cards.nine.processes.messages.InstallationsMessages._
 import cards.nine.processes.messages.UserMessages._
 import cards.nine.processes.messages.rankings.GetRankedDeviceApps._
@@ -32,68 +30,6 @@ object Converters {
       sessionToken = response.sessionToken
     )
 
-  implicit def toApiCreateOrUpdateCollectionResponse(
-    response: CreateOrUpdateCollectionResponse
-  ): ApiCreateOrUpdateCollectionResponse =
-    ApiCreateOrUpdateCollectionResponse(
-      publicIdentifier = response.publicIdentifier,
-      packagesStats    = response.packagesStats
-    )
-
-  def toCreateCollectionRequest(
-    request: ApiCreateCollectionRequest,
-    collectionInfo: NewSharedCollectionInfo,
-    userContext: UserContext
-  ): CreateCollectionRequest =
-    CreateCollectionRequest(
-      collection = SharedCollectionData(
-        publicIdentifier = collectionInfo.identifier.value,
-        userId           = Option(userContext.userId.value),
-        publishedOn      = collectionInfo.currentDate.value,
-        author           = request.author,
-        name             = request.name,
-        views            = request.views,
-        category         = request.category,
-        icon             = request.icon,
-        community        = request.community,
-        packages         = request.packages
-      )
-    )
-
-  def toApiIncreaseViewsCountByOneResponse(
-    response: IncreaseViewsCountByOneResponse
-  ): ApiIncreaseViewsCountByOneResponse =
-    ApiIncreaseViewsCountByOneResponse(
-      publicIdentifier = response.publicIdentifier
-    )
-
-  def toApiSharedCollection[A](info: SharedCollectionWithAppsInfo[A])(toApiApp: A ⇒ ApiCollectionApp): ApiSharedCollection =
-    ApiSharedCollection(
-      publicIdentifier = info.collection.publicIdentifier,
-      publishedOn      = info.collection.publishedOn,
-      author           = info.collection.author,
-      name             = info.collection.name,
-      views            = info.collection.views,
-      category         = info.collection.category,
-      icon             = info.collection.icon,
-      community        = info.collection.community,
-      owned            = info.collection.owned,
-      packages         = info.collection.packages,
-      appsInfo         = info.appsInfo map toApiApp,
-      subscriptions    = info.collection.subscriptionsCount
-    )
-
-  def toApiCollectionApp(card: FullCard): ApiCollectionApp =
-    ApiCollectionApp(
-      packageName = card.packageName,
-      title       = card.title,
-      free        = card.free,
-      icon        = card.icon,
-      stars       = card.stars,
-      downloads   = card.downloads,
-      categories  = card.categories
-    )
-
   def toFullCard(packageName: Package, apiDetails: ApiSetAppInfoRequest): FullCard =
     FullCard(
       packageName = packageName,
@@ -105,20 +41,6 @@ object Converters {
       categories  = apiDetails.categories,
       screenshots = apiDetails.screenshots
     )
-
-  def toApiCollectionApp(card: BasicCard): ApiCollectionApp =
-    ApiCollectionApp(
-      packageName = card.packageName,
-      title       = card.title,
-      free        = card.free,
-      icon        = card.icon,
-      stars       = card.stars,
-      downloads   = card.downloads,
-      categories  = Nil
-    )
-
-  def toApiSharedCollectionList(response: GetCollectionsResponse): ApiSharedCollectionList =
-    ApiSharedCollectionList(response.collections map { col ⇒ toApiSharedCollection(col)(toApiCollectionApp) })
 
   def toUpdateInstallationRequest(
     request: ApiUpdateInstallationRequest,
@@ -137,12 +59,6 @@ object Converters {
       androidId   = response.androidId,
       deviceToken = response.deviceToken
     )
-
-  def toApiSubscribeResponse(response: SubscribeResponse): ApiSubscribeResponse =
-    ApiSubscribeResponse()
-
-  def toApiUnsubscribeResponse(response: UnsubscribeResponse): ApiUnsubscribeResponse =
-    ApiUnsubscribeResponse()
 
   def toMarketAuth(googlePlayContext: GooglePlayContext, userContext: UserContext): MarketCredentials =
     MarketCredentials(
@@ -201,11 +117,6 @@ object Converters {
       stars       = card.stars,
       downloads   = card.downloads,
       screenshots = Nil
-    )
-
-  def toApiGetSubscriptionsByUser(response: GetSubscriptionsByUserResponse): ApiGetSubscriptionsByUser =
-    ApiGetSubscriptionsByUser(
-      subscriptions = response.subscriptions
     )
 
   def toApiGetRecommendationsResponse(response: CardList[FullCard]): ApiGetRecommendationsResponse =
